requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
PyJWT==2.9.0
pytest==8.2.1

flask-restx~=1.3.0